Remove added dependencies for serverless
